added theos components, made averagely compilable (with in-house framework headers)

+++ Symbolicator.mm
@@ -6,7 +6,7 @@
 //
 //
 
-#import <AppSupport/AppSupport.h>
+#import "AppSupport/AppSupport.h"
 
 #include <mach/mach.h>
 #include <objc/runtime.h>
@@ -27,7 +27,7 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
     entry->name[strlen(className)+2] = ' ';
     strcpy(entry->name+strlen(className)+3,methodName);
     strcpy(entry->name+strlen(className)+strlen(methodName)+3,"]");
-    
+
     return entry;
 }
 
@@ -37,15 +37,15 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
 {
     self = [super init];
     if (self) {
-        
+
         _symbolicator = [[VMUSymbolicator symbolicatorForTask:mach_task_self()] retain];
         _processInfo = [[VMUProcessDescription alloc] initWithPid:0 orTask:mach_task_self()];
         _machContainer = [[VMUMachTaskContainer machTaskContainerWithTask:mach_task_self()] retain];
-        
+
         dispatch_async(dispatch_get_global_queue( DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
             [self load];
         });
-        
+
     }
     return self;
 }
@@ -53,68 +53,68 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
 - (void) load
 {
     NSLog(@"Symbolicator: Loading all methods addresses from objc runtime");
-    
+
     int numClasses;
     unsigned int numInstanceMethods;
     unsigned int numClassMethods;
     Class * classes = NULL;
     Method * instanceMethods = NULL;
     Method * classMethods = NULL;
-    
+
     classes = NULL;
     char className[128];
     numClasses = objc_getClassList(NULL, 0);
-        
+
     unsigned int imp;
-    
+
     _maxAddress = 0;
     _minAddress = 2^32;
-    
+
     timeval time1, time2;
     gettimeofday(&time1, NULL);
     long millis1 = (time1.tv_sec * 1000) + (time1.tv_usec / 1000);
-    
+
     MethodEntry *entry;
     if (numClasses > 0)
     {
         classes = (Class *)malloc(sizeof(Class) * numClasses);
         numClasses = objc_getClassList(classes, numClasses);
-        
+
         for (int i=0;i<numClasses;i++)
         {
             classMethods = class_copyMethodList(objc_getMetaClass(class_getName(classes[i])), &numClassMethods);
             instanceMethods = class_copyMethodList(classes[i], &numInstanceMethods);
-            
+
             for (int j=0; j<numClassMethods; j++) {
-                imp = (unsigned int)method_getImplementation(classMethods[j]);
+                imp = (NSUInteger)method_getImplementation(classMethods[j]);
                 entry = createMethodEntry(imp, class_getName(classes[i]), sel_getName(method_getName(classMethods[j])), YES);
-                
+
                 HASH_ADD_INT(_methodList, imp, entry);
-                
+
                 if (imp > _maxAddress) _maxAddress = imp;
                 if (imp < _minAddress) _minAddress = imp;
             }
-            
+
             for (int j=0; j<numInstanceMethods; j++) {
-                imp = (unsigned int)method_getImplementation(instanceMethods[j]);
+                imp = (NSUInteger)method_getImplementation(instanceMethods[j]);
                 entry = createMethodEntry(imp, class_getName(classes[i]), sel_getName(method_getName(instanceMethods[j])), NO);
-                
+
                 HASH_ADD_INT(_methodList, imp, entry);
-                
+
                 if (imp > _maxAddress) _maxAddress = imp;
                 if (imp < _minAddress) _minAddress = imp;
             }
-            
+
             free(classMethods);
             free(instanceMethods);
         }
-        
+
         free(classes);
     }
-    
+
     gettimeofday(&time2, NULL);
     long millis2 = (time2.tv_sec * 1000) + (time2.tv_usec / 1000);
-    
+
     NSLog(@"Symbolicator: All addresses loaded and sorted %ld milliseconds",millis2-millis1);//-millis1/1000.0);
 
 }
@@ -122,33 +122,33 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
 - (NSDictionary *)symbolicateAddresses:(NSArray *)addresses
 {
     NSMutableDictionary *response = [NSMutableDictionary dictionaryWithCapacity:addresses.count];
-    
+
     NSString *hexAddress;
-    
+
     NSString *methodName;
     unsigned long long longLongAddress;
-    
+
     // Try to get the symbols for all requested addresses
     for (NSNumber *address in addresses)
     {
         longLongAddress = [address unsignedLongLongValue];
         hexAddress = [NSString stringWithFormat:@"0x%08llx",longLongAddress];
 
-        
+
         methodName = [self findMethod:address slide:[self slideForAddress:longLongAddress]];
         if (methodName)
         {
             response[hexAddress] = methodName;
             continue;
         }
-        
+
         VMUSymbol *symbol = [_symbolicator symbolForAddress:longLongAddress];
 
         if ([symbol name])
             response[hexAddress] = [symbol name];
-        
+
     }
-    
+
     return response;
 }
 
@@ -156,26 +156,26 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
 {
     unsigned int newAddress = [address unsignedIntValue];
     unsigned searchCount = 0;
-    
+
     struct MethodEntry *entry;
-    
+
     do {
-        
+
         entry = NULL;
         HASH_FIND_INT( _methodList, &newAddress, entry );
-        
+
         if (entry)
             return [NSString stringWithFormat:@"(0x%08x): %s", newAddress-slide-1, entry->name];
-        
+
         searchCount++;
-        
+
         if (searchCount > 8096)
             return [NSString stringWithFormat:@"(0x%08x):", [address unsignedIntValue]-slide-1];
-        
+
         newAddress = newAddress-1;
-        
+
     } while (newAddress!=_minAddress);
-    
+
     return nil;
 }
 
@@ -185,19 +185,19 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
     VMUMachOHeader *header;
     VMURange range;
     unsigned long long textStart;
-    
+
     for (NSDictionary *image in _processInfo.binaryImages)
     {
         range.location = [image[@"StartAddress"] unsignedLongLongValue];
         range.length = [image[@"Size"] unsignedLongLongValue];
-        
+
         if (address >= range.location &&
             address <= range.location+range.length)
         {
             memory = [VMUMemory_NonContiguousTask memoryWithMachTaskContainer:_machContainer addressRange:range architecture:[VMUArchitecture armArchitecture]];
             header = [VMUMachOHeader headerWithMemory:memory address:range.location name:image[@"Identifier"] path:image[@"ExecutablePath"] timestamp:[NSDate date]];
             textStart = [[header segmentNamed:@"__TEXT"] vmaddr];
-            
+
             return range.location-textStart;
         }
     }
@@ -222,9 +222,9 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
     self = [super init];
     if (self) {
         _descriptions = [[NSMutableArray alloc] initWithCapacity:[array count]];
-        
+
         NSString *pos, *nextpos;
-        
+
         for (int i=0;i<[array count];i++) //fugly stuff - must eliminate first line (that's Symbolicator's call you don't need) and preserve the numbers with the correct spacing
         {
 //            pos = [array[i] componentsSeparatedByString:@" "][0];
@@ -234,24 +234,24 @@ MethodEntry *createMethodEntry(unsigned imp, const char *className,const char *m
             _descriptions[i] = array[i];
         }
     }
-    
+
     return self;
 }
 
-- (unsigned int)count
+- (NSUInteger)count
 {
     return [_descriptions count];
 }
 
-- (id)objectAtIndex:(unsigned int)index
+- (id)objectAtIndex:(NSUInteger)index
 {
     return _descriptions[index];
 }
 
-- (id)descriptionWithLocale:(id)locale indent:(unsigned int)level
+- (id)descriptionWithLocale:(id)locale indent:(NSUInteger)level
 {
     NSString *indentation = [@"" stringByPaddingToLength:level withString:@" " startingAtIndex:0];
-    
+
     NSMutableString *output = [NSMutableString stringWithCapacity:4096];
     [output appendString:indentation];
     [output appendString:@"(\n"];
